Suppress GCC warnings

+++ source/core/coding/coding_units.hpp
@@ -117,7 +117,7 @@ class j2k_codeblock : public j2k_region {
   [[nodiscard]] uint8_t get_Mb() const;
   uint8_t *get_compressed_data();
   void set_compressed_data(uint8_t *buf, uint16_t size, uint16_t Lref = 0);
-  void create_compressed_buffer(buf_chain *tile_buf, uint16_t buf_limit, const uint16_t &layer);
+  void create_compressed_buffer(buf_chain *tile_buf, int32_t buf_limit, const uint16_t &layer);
   void update_sample(const uint8_t &symbol, const uint8_t &p, const int16_t &j1, const int16_t &j2) const;
   void update_sign(const int8_t &val, const int16_t &j1, const int16_t &j2) const;
   [[nodiscard]] uint8_t get_sign(const int16_t &j1, const int16_t &j2) const;
