+++ source/core/coding/ht_block_encoding.cpp
@@ -112,8 +112,8 @@ void j2k_codeblock::set_MagSgn_and_sigma(uint32_t &or_val) {
     auto vabsmask  = _mm256_set1_epi16((int16_t)0x7FFF);
     auto vzero     = _mm256_setzero_si256();
     // simd
-    uint16_t simdlen = round_down(this->size.x, 16U);
-    for (uint16_t j = 0; j < simdlen; j += 16) {
+    int32_t simdlen = round_down(this->size.x, 16);
+    for (int32_t j = 0; j < simdlen; j += 16) {
       auto coeff16 = _mm256_loadu_si256((__m256i *)(sp + j));
       // auto vsmag   = _mm256_and_si256(coeff16, vpLSB);
       auto vsign = _mm256_srli_epi16(_mm256_and_si256(coeff16, vsignmask), 15);
@@ -142,11 +142,11 @@ void j2k_codeblock::set_MagSgn_and_sigma(uint32_t &or_val) {
     }
     or_val |= !_mm256_testz_si256(vzero, vabsmask);
     // remaining
-    for (uint16_t j = simdlen; j < this->size.x; ++j) {
+    for (int32_t j = simdlen; j < static_cast<int32_t>(this->size.x); ++j) {
       int32_t temp  = sp[j];
       uint32_t sign = (static_cast<uint32_t>(temp) & 0x80000000) >> 31;
-      dstblk[j] |= (temp & pLSB) << SHIFT_SMAG;
-      dstblk[j] |= (sign) << SHIFT_SSGN;
+      dstblk[j] |= static_cast<uint8_t>((temp & pLSB) << SHIFT_SMAG);
+      dstblk[j] |= static_cast<uint8_t>((sign) << SHIFT_SSGN);
       temp = (temp < 0) ? -temp : temp;
       temp &= 0x7FFFFFFF;
       temp >>= pshift;
